requirements.txt
python-telegram-bot==21.6
groq==0.11.0
mistralai==1.2.5
python-dotenv==1.0.1
